# Varuint optimization