--- a/benches/read_many_structs.rs
+++ b/benches/read_many_structs.rs
@@ -2,17 +2,17 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use ion_rs::lazy::decoder::LazyDecoder;
 use ion_rs::lazy::encoding::TextEncoding_1_1;
 use ion_rs::lazy::r#struct::LazyStruct;
-use ion_rs::lazy::reader::{LazyApplicationReader, LazyTextReader_1_1};
+use ion_rs::lazy::reader::{LazyApplicationReader, LazyBinaryReader, LazyTextReader_1_1};
 use ion_rs::lazy::value::LazyValue;
 use ion_rs::lazy::value_ref::ValueRef;
 use ion_rs::{Element, Format, IonResult, TextKind};
 use ion_rs::{ElementReader, IonData};
 
-fn rewrite_as_compact_text(pretty_ion: &str) -> IonResult<String> {
+fn rewrite_as(pretty_ion: &str, format: Format) -> IonResult<Vec<u8>> {
     let values = Element::read_all(pretty_ion).unwrap();
     let mut buffer = Vec::new();
-    Element::write_all_as(&values, Format::Text(TextKind::Compact), &mut buffer)?;
-    Ok(String::from_utf8(buffer).unwrap())
+    Element::write_all_as(&values, format, &mut buffer)?;
+    Ok(buffer)
 }
 
 pub fn criterion_benchmark(c: &mut Criterion) {
@@ -26,7 +26,8 @@ pub fn criterion_benchmark(c: &mut Criterion) {
             'format': "Request status: {} Client ID: {} Client Host: {} Client Region: {} Timestamp: {}",
             'parameters': ["SUCCESS","example-client-1","aws-us-east-5f-18b4fa","region 4","2022-12-07T20:59:59.744000Z",],
         }"#.repeat(NUM_VALUES);
-    let data_1_0 = rewrite_as_compact_text(&pretty_data_1_0).unwrap();
+    let text_1_0_data = rewrite_as(&pretty_data_1_0, Format::Text(TextKind::Compact)).unwrap();
+    let binary_1_0_data = rewrite_as(&pretty_data_1_0, Format::Binary).unwrap();
     let template_text = r#"
             (macro event (timestamp thread_id thread_name client_num host_id parameters)
                 {
@@ -46,18 +47,19 @@ pub fn criterion_benchmark(c: &mut Criterion) {
             )
         "#;
 
-    let data_1_1 = r#"(:event 1670446800245 418 "6" "1" "18b4fa" (:values "region 4" "2022-12-07T20:59:59.744000Z"))"#.repeat(NUM_VALUES);
+    let text_1_1_data = r#"(:event 1670446800245 418 "6" "1" "18b4fa" (:values "region 4" "2022-12-07T20:59:59.744000Z"))"#.repeat(NUM_VALUES);
 
-    println!("Ion 1.0 data size: {} bytes", data_1_0.len());
-    println!("Ion 1.1 data size: {} bytes", data_1_1.len());
+    println!("Bin  Ion 1.0 data size: {} bytes", binary_1_0_data.len());
+    println!("Text Ion 1.0 data size: {} bytes", text_1_0_data.len());
+    println!("Text Ion 1.1 data size: {} bytes", text_1_1_data.len());
 
     // As a sanity check, materialize the data from both the Ion 1.0 and 1.1 streams and make sure
     // that they are equivalent before we start measuring the time needed to read them.
-    let seq_1_0 = LazyTextReader_1_1::new(data_1_0.as_bytes())
+    let seq_1_0 = LazyTextReader_1_1::new(&text_1_0_data)
         .unwrap()
         .read_all_elements()
         .unwrap();
-    let mut reader_1_1 = LazyTextReader_1_1::new(data_1_1.as_bytes()).unwrap();
+    let mut reader_1_1 = LazyTextReader_1_1::new(text_1_1_data.as_bytes()).unwrap();
     reader_1_1.register_template(template_text).unwrap();
     let seq_1_1 = reader_1_1.read_all_elements().unwrap();
     assert!(
@@ -97,10 +99,28 @@ pub fn criterion_benchmark(c: &mut Criterion) {
         Ok(count)
     }
 
+    c.bench_function("binary 1.0: scan all", |b| {
+        b.iter(|| {
+            let mut reader = LazyBinaryReader::new(&binary_1_0_data).unwrap();
+            while let Some(item) = reader.next().unwrap() {
+                black_box(item);
+            }
+        })
+    });
+    c.bench_function("binary 1.0: read all", |b| {
+        b.iter(|| {
+            let mut reader = LazyBinaryReader::new(&binary_1_0_data).unwrap();
+            let mut num_values = 0usize;
+            while let Some(item) = reader.next().unwrap() {
+                num_values += count_value_and_children(&item).unwrap();
+            }
+            let _ = black_box(num_values);
+        })
+    });
     c.bench_function("text 1.0: scan all", |b| {
         b.iter(|| {
             let mut reader =
-                LazyApplicationReader::<'_, TextEncoding_1_1>::new(data_1_0.as_bytes()).unwrap();
+                LazyApplicationReader::<'_, TextEncoding_1_1>::new(&text_1_0_data).unwrap();
             while let Some(item) = reader.next().unwrap() {
                 black_box(item);
             }
@@ -109,7 +129,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     c.bench_function("text 1.0: read all", |b| {
         b.iter(|| {
             let mut reader =
-                LazyApplicationReader::<'_, TextEncoding_1_1>::new(data_1_0.as_bytes()).unwrap();
+                LazyApplicationReader::<'_, TextEncoding_1_1>::new(&text_1_0_data).unwrap();
             let mut num_values = 0usize;
             while let Some(item) = reader.next().unwrap() {
                 num_values += count_value_and_children(&item).unwrap();
@@ -120,7 +140,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     c.bench_function("text 1.0: read 'format' field", |b| {
         b.iter(|| {
             let mut reader =
-                LazyApplicationReader::<'_, TextEncoding_1_1>::new(data_1_0.as_bytes()).unwrap();
+                LazyApplicationReader::<'_, TextEncoding_1_1>::new(&text_1_0_data).unwrap();
             let mut num_values = 0usize;
             while let Some(value) = reader.next().unwrap() {
                 let s = value.read().unwrap().expect_struct().unwrap();
@@ -132,7 +152,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     });
     c.bench_function("text 1.1: scan all", |b| {
         b.iter(|| {
-            let mut reader = LazyTextReader_1_1::new(data_1_1.as_bytes()).unwrap();
+            let mut reader = LazyTextReader_1_1::new(text_1_1_data.as_bytes()).unwrap();
             reader.register_template(template_text).unwrap();
             while let Some(item) = reader.next().unwrap() {
                 black_box(item);
@@ -141,7 +161,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     });
     c.bench_function("text 1.1: read all", |b| {
         b.iter(|| {
-            let mut reader = LazyTextReader_1_1::new(data_1_1.as_bytes()).unwrap();
+            let mut reader = LazyTextReader_1_1::new(text_1_1_data.as_bytes()).unwrap();
             reader.register_template(template_text).unwrap();
             let mut num_values = 0usize;
             while let Some(item) = reader.next().unwrap() {
@@ -153,7 +173,8 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     c.bench_function("text 1.1: read 'format' field", |b| {
         b.iter(|| {
             let mut reader =
-                LazyApplicationReader::<'_, TextEncoding_1_1>::new(data_1_1.as_bytes()).unwrap();
+                LazyApplicationReader::<'_, TextEncoding_1_1>::new(text_1_1_data.as_bytes())
+                    .unwrap();
             reader.register_template(template_text).unwrap();
             let mut num_values = 0usize;
             while let Some(value) = reader.next().unwrap() {