Optimizes reading `VarUInt`s that are 1 or 2 bytes long. Also adds a binary 1.0 flavor of the existing `read_many_structs` benchmark.

| Benchmark | Before | After |
|:----|:---:| :---:|
| Binary 1.0/scan all | 843.74 us | 776.22 us |
| Binary 1.0/read all | 16.366 ms | 15.391 ms | 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.